feat(load.cc): Load general tree-like cache structures
The `loading_tree` structure manages the binary cache tree. This
design allows arbitrary binary trees to be loaded from cache. The
loading is done via recursion over both children.

The nodes are referenced via indices and their relationships are
described by the parents_, left_, and right_ vectors.
The first node (idx=0) is always the default timetable. After that
follow the other source based nodes, and then the intermediate merging
nodes. In total, there are always 2(n+1)-1 nodes, where n is the number
of timetable sources.

Each node has a corresponding cache location to which it will save its
result. By default, it is the node index, but that can be changed if
necessary. To avoid issues between the automatic and the manual indices,
the manual indices can't be set to an automatic value.

Building up the tree has 5 steps:
 - Insert the default node at index 0
 - Insert source nodes up to n+1
 - Determine the updated sources by comparing with the previous cache
 - Insert intermediate nodes. They are updated if either child is updated
 - Write the cache structure to disk. From now on, it is constant.

Loading from the tree-like cache is done via tree traversal. There are
3 types of nodes:
 - The default timetable node at index 0,
 - the loading nodes up to l.source_paths_.size(),
 - the merging nodes after that.

1. The default timetable gets constructed on the fly.
2. The latter two will check whether they need to be recomputed. If they
   don't, they'll try to read the timetable from the cache.
3. If the node needs to be recomputed (either due to update or cache
   failure), we need to determine which kind of node it is:
3.1. Merging nodes call both children recursively. If that fails and they
     return nothing, then they get skipped. If both have been loaded
     successfully, their timetables are merged and written. The shapes
     only get saved if both are available; Usually either both are or both
     are not, as that is dependent on the global setting of whether to
     save the shapes or not. If they get saved, the left gets added first,
     then the right one.
3.2. If it is a source based node, then it will try to load from source.
     On success, it writes the resulting timetable into the cache
     location; The shapes are automatically written to there during the
     loading from source.
     In case the loading fails, this is propagated upwards, where it
     can be skipped.

Finally, the timetable gets finalized. This is unchanged.