feat(load.cc): Complete Binary Tree Cache Structure
The current tree structure is close to a degenerate binary tree.
Using a complete binary tree is usually more efficient for evenly
spread out updates and is easily computable.

The initial slots are filled with the source based nodes. After
that, the merging nodes are created in a similar manner to binary
heaps. The most obvious difference is the reversed
ordering due to insertion order (starting at the leaves instead of
the root). This algorithm produces a complete binary tree.

The algorithm chooses the two left-most nodes without a parent to
set as children for the next node. This is done until the next node
would be only node left.

This means that the merge order is not necessarily the same as the
ordering inside the array if the source nodes are at different depths
(always the case unless the number of leafs is a power of 2). However,
the default timetable node is required to be the left-most merged
node due to the special stations.
With this tree creation algorithm, it suffices to rotate the sources
by the difference to the next power of 2 (that is the number of leafs
missing on the last level).