feat(load.cc): Node indices are independent from the source indices.
Some nodes are source based and thus need access to their underlying
source. Previously, this was handled by fixing the positions relative
to each other. Since the nodes may need to be reordered to facilitate
the tree creation, storing the mapping explicitly makes that easier.

The addition of source based nodes is also now handled inside the
loading_tree, only requiring a simple function call to add a source.