feat(load.cc): Dynamically match cached nodes
Currently, the cache strategy assumes that the cached tree structure
is identical to the current loading tree. Any deviation leads to
cache invalidation and wrong results.
This puts severe restrictions not only on the cache structures that
can be accepted, but also on the number of nodes contained within
them.

In particular, adding or removing sources invalidates the cache
even if the same algorithm is used, leading to extreme performance
degradation.

The dynamic matching of nodes during the tree building checks that
the assumptions are correct and attempts to make good use of the
existing cache. The approach here is not optimal for all existing
tree structures, and doesn't try to adjust the new cache structure
to the existing one. It does, however, produce correct results with
all valid cache tree structures.

---

The matching works on this algorithm:

A merge node can be reused if the underlying sources are unchanged.
The ordering of the sources doesn't need to be checked because any
valid reordering produces essentially the same timetable. Thus,
the sources can be sorted for easier matching. Due to the tree
structure, the sets of children are unique for each node.

Source based nodes are the same if the source data is unchanged.
Thus, the path can be used to quickly determine a match candidate.

The default timetable node has the reserved cache index of 0.

When a match has been found, it's cache index is transferred to the
new node. That index might already be used by another node, in which
case they swap them. Since each cache index is unique, this swap
doesn't mess with the matching.

---

In summary, the loading from source is only done with new sources,
while cached merging nodes may or may not be useful depending on
the differences in structure.
Best performance is achieved when prepending sources on the left.